Basics

Basics

The package exports the following functions for parameters identifiability analysis, confidence intervals evaluation and results visualization.

Functions

LikelihoodProfiler.params_intervals — Function.

params_intervals(init_params::Vector{Float64}, id::Int,
loss_crit::Float64, loss_func::Function; <keyword arguments>)

Computes confidence interval for id parameter of init_params vector and loss_func according to loss_crit confidence level.

Returns ParamInterval structure storing all input data and estimated confidence interval.

Arguments

method::Symbol: computational method (:ONE_PASS,:D2D_PLE).

logscale_all::Bool: set logscale for all parameters to true / false (default false).

logscale::Vector{Bool}: set logscale for each parameter (default false for all parameters).

scan_bound::Vector{Float64}: search bounds for id parameter (default [-9.,9.]).

local_alg::Symbol: fitting algorithm (default :LN_NELDERMEAD).

bounds::Vector{Vector{Float64}}: bound constraints for all parameters (default [-Inf,Inf]).

max_iter::Int: maximum loss_func evaluations (default 10^5).

ptol::Float64: fitting tolerance for optimizer (default 1e-3).

losstol::Float64: constraints tolerance (default 1e-3).

source

LikelihoodProfiler.params_plot — Function.

params_plot(params::Vector{Float64}, id::Int, loss_func::Function,
interval::Tuple{Float64,Float64}; <keyword arguments>)

Computes adapted_grid for loss_func and id parameter values from the interval. See also: Plots.adapted_grid

Arguments

fit_alg::Symbol: fitting algorithm (default :LN_NELDERMEAD).

bounds::Vector{Vector{Float64}}: bound constraints for all parameters (default [-Inf,Inf]).

tol::Float64: fitting tolerance (default ftol_abs = 1e-3).

max_recursions::Int: how many times each interval is allowed to be refined (default 2).

source