Solution interface

ProfileLikelihoodSolution type is designed to contain the results of a profile likelihood analysis.

LikelihoodProfiler.ProfileLikelihoodSolution — Type

ProfileLikelihoodSolution{probType,P}

Contains the results of a profile likelihood analysis.

Fields

prob::probType: The profile likelihood problem ProfileLikelihoodProblem.

profiles::P: The computed profiles.

elapsed_time::Float64: The time elapsed during the computation.

Selectors

A number of selectors are available to extract information from the sol::ProfileLikelihoodSolution object. These can be applied to each computed profile sol[i]:

get_endpoints(sol[i]): Returns the confidence interval (CI) endpoints, marking the intersection of the profile with the threshold.

get_retcodes(sol[i]): Returns the retcodes of the CI endpoints estimation.

get_stats(sol[i]): Returns the statistics of the profile computation.

Retcodes

sol::ProfileLikelihoodSolution outputs the following retcodes, which are accessible with get_retcodes(sol[i]) function:

:Identifiable - the profile has intersection with the predefined threshold.

:NonIdentifiable - the profile doesn't intersect the predefined threshold.

:MaxIters - maximum number of iterations reached while computing the profile. See maxiters argument to the solve.

:Failure - the solver (optimizer or integrator) reported failure status, profiling was interrupted.

Visualization and tabular representation

The recipes are defined to visualize profiles saved in sol::ProfileLikelihoodSolution with Plots.jl package: plot(sol), plot(sol[i]). The following keyword arguments can be used in plot function:

steps::Bool - whether to scatter steps performed by the profiler. Defaults to true.

threshold::Bool - whether to plot threshold defined in ProfileLikelihoodProblem. Defaults to isfinite(threshold)

Also each profile contained in the sol::ProfileLikelihoodSolution can be represented as a DataFrame with DataFrame(sol[i]).